Start early fix for foreseen parametric activations
The activation functions can also be treated as units, with
numInputs==numOutputs.
This will allow parametric activation functions to have their parameter
as part of their model.

This commit is the start of the fix, with initial work to create
summation units.

The combination of a summation unit and an activation unit will produce
a perceptron. Whether a perceptron wrapper will be created is another
matter.

This should also lay the groundwork for RBF units: combination of a
"distance" unit and an activation unit.